requirements.txt
anthropic>=0.52.0
schedule
python-dotenv
praw
requests
